requirements.txt
# requirements.txt
python-telegram-bot
SQLAlchemy
python-dotenv
httpx
pg8000